# [mono][wasm] Allow AOTing methods with catch/filter clauses.