--- a/src/mono/mono/mini/interp/interp-internals.h
+++ b/src/mono/mono/mini/interp/interp-internals.h
@@ -210,6 +210,7 @@ typedef struct {
 	/* Lets interpreter know it has to resume execution after EH */
 	gboolean has_resume_state;
 	/* Frame to resume execution at */
+	/* Can be NULL if the exception is caught in an AOTed frame */
 	InterpFrame *handler_frame;
 	/* IP to resume execution at */
 	const guint16 *handler_ip;
@@ -231,8 +232,6 @@ typedef struct {
 	guchar *stack_pointer;
 	/* Used for allocation of localloc regions */
 	FrameDataAllocator data_stack;
-	/* Whenever to unwind from interpreter code back to the caller */
-	gboolean pending_unwind;
 } ThreadContext;
 
 typedef struct {