… src/libraries/System.Text.Json/tests/System.Text.Json.SourceGeneration.Tests/System.Text.Json.SourceGeneration.Roslyn3.11.Tests.csproj
… src/libraries/System.Text.Json/tests/System.Text.Json.SourceGeneration.Tests/System.Text.Json.SourceGeneration.Roslyn4.0.Tests.csproj
… src/libraries/System.Text.Json/tests/System.Text.Json.Tests/System.Text.Json.Tests.csproj
… src/mono/mono/mini/mini-llvm.c
… src/mono/mono/mini/mini-runtime.c
… src/mono/mono/mini/mini-runtime.h
… src/mono/mono/mini/mini.c
… src/mono/mono/mini/mini.h
… src/mono/sample/wasm/browser-bench/Console/Makefile
… src/mono/sample/wasm/browser-bench/Console/Wasm.Console.Bench.Sample.csproj