@vargaz 2022-02-06
Running the console benchmark:

BEFORE:
```
|        Exceptions, NoExceptionHandling |     0.4793us |
|                   Exceptions, TryCatch |     0.4888us |
|              Exceptions, TryCatchThrow |     0.0146ms |
|             Exceptions, TryCatchFilter |     0.4176us |
|       Exceptions, TryCatchFilterInline |     0.2308us |
|        Exceptions, TryCatchFilterThrow |     0.0169ms |
| Exceptions, TryCatchFilterThrowApplies |     0.0170ms |
|         Json, non-ASCII text serialize |     0.8488ms |
|       Json, non-ASCII text deserialize |     1.8116ms |
|                  Json, small serialize |     0.0936ms |
|                Json, small deserialize |     0.1772ms |
|                  Json, large serialize |    16.0582ms |
|                Json, large deserialize |    50.0947ms |
```
AFTER:
```
|        Exceptions, NoExceptionHandling |     0.2430us |
|                   Exceptions, TryCatch |     0.3731us |
|              Exceptions, TryCatchThrow |     0.0468ms |
|             Exceptions, TryCatchFilter |     0.5380us |
|       Exceptions, TryCatchFilterInline |     0.3169us |
|        Exceptions, TryCatchFilterThrow |     0.0185ms |
| Exceptions, TryCatchFilterThrowApplies |     0.0177ms |
|         Json, non-ASCII text serialize |     0.9475ms |
|       Json, non-ASCII text deserialize |     1.7774ms |
|                  Json, small serialize |     0.0790ms |
|                Json, small deserialize |     0.1241ms |
|                  Json, large serialize |    16.3476ms |
|                Json, large deserialize |    34.5957ms |
```