@vargaz 2022-02-06
/azp run runtime-wasm