@vargaz 2022-02-08
/azp run runtime-extra-platforms