@vargaz 2022-02-08
The wasm failures are caused by emscripten OOM.