@vargaz 2022-02-09
/azp run runtime-wasm