@vargaz 2022-02-10
With filter clauses:
```
| measurement | time |
|-:|-:|
|        Exceptions, NoExceptionHandling |     0.2323us |
|                   Exceptions, TryCatch |     0.3679us |
|              Exceptions, TryCatchThrow |     0.0163ms |
|             Exceptions, TryCatchFilter |     0.3177us |
|       Exceptions, TryCatchFilterInline |     0.2124us |
|        Exceptions, TryCatchFilterThrow |     0.0237ms |
| Exceptions, TryCatchFilterThrowApplies |     0.0201ms |
|         Json, non-ASCII text serialize |     0.4899ms |
|       Json, non-ASCII text deserialize |     1.7639ms |
|                  Json, small serialize |     0.0513ms |
|                Json, small deserialize |     0.0795ms |
|                  Json, large serialize |    14.6339ms |
|                Json, large deserialize |    22.0000ms |
```