@vargaz 2022-02-11
/azp run runtime-wasm