@vargaz 2022-02-13
/azp run runtime-wasm