@vargaz 2022-02-15
/azp run runtime-wasm