@vargaz 2022-02-16
/azp run runtime-wasm