@vargaz 2022-02-17
/azp run runtime-wasm