@vargaz 2022-02-27
/azp run runtime-extra-platforms