@vargaz 2022-02-28
/azp run runtime-wasm