@vargaz 2022-02-28
/azp run runtime-extra-platforms